Refactor proxies
