+++ packages/backend/src/core/discovery/analyzeItem.ts
@@ -1,11 +1,11 @@
 import { EthereumAddress } from '@l2beat/types'
 
-import { detectProxy } from './detectProxy'
 import { DiscoveryConfig } from './DiscoveryConfig'
 import { getMetadata } from './getMetadata'
 import { executeHandlers } from './handlers/executeHandlers'
 import { getHandlers } from './handlers/getHandlers'
 import { DiscoveryProvider } from './provider/DiscoveryProvider'
+import { detectProxy } from './proxies'
 import { ContractParameters, ContractValue } from './types'
 
 export interface AnalyzedData extends ContractParameters {
