+++ packages/backend/src/core/discovery/proxies/ArbitrumProxy.ts
@@ -2,7 +2,7 @@ import { Bytes, EthereumAddress } from '@l2beat/types'
 
 import { DiscoveryProvider } from '../provider/DiscoveryProvider'
 import { bytes32ToAddress } from '../utils/address'
-import { Eip1967Proxy } from './Eip1967Proxy'
+import { getAdmin, getImplementation } from './Eip1967Proxy'
 import { ProxyDetection } from './types'
 
 // keccak256('eip1967.proxy.implementation.secondary') - 1)
@@ -10,7 +10,7 @@ const SECONDARY_IMPLEMENTATION_SLOT = Bytes.fromHex(
   '0x2b1dbce74324248c222f0ec2d5ed7bd323cfc425b336f0253c5ccfda7265546d',
 )
 
-export async function getSecondaryImplementation(
+async function getSecondaryImplementation(
   provider: DiscoveryProvider,
   address: EthereumAddress,
 ) {
@@ -19,7 +19,7 @@ export async function getSecondaryImplementation(
   )
 }
 
-async function detect(
+export async function detectArbitrumProxy(
   provider: DiscoveryProvider,
   address: EthereumAddress,
 ): Promise<ProxyDetection | undefined> {
@@ -28,8 +28,8 @@ async function detect(
     return
   }
   const [adminImplementation, admin] = await Promise.all([
-    Eip1967Proxy.getImplementation(provider, address),
-    Eip1967Proxy.getAdmin(provider, address),
+    getImplementation(provider, address),
+    getAdmin(provider, address),
   ])
   return {
     implementations: [adminImplementation, userImplementation],
@@ -41,11 +41,4 @@ async function detect(
       userImplementation,
     },
   }
-}
-
-export const ArbitrumProxy = {
-  getAdmin: Eip1967Proxy.getAdmin,
-  getAdminImplementation: Eip1967Proxy.getImplementation,
-  getUserImplementation: getSecondaryImplementation,
-  detect,
 }
