+++ packages/backend/src/core/discovery/proxies/Eip1967Proxy.ts
@@ -9,7 +9,7 @@ const IMPLEMENTATION_SLOT = Bytes.fromHex(
   '0x360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc',
 )
 
-async function getImplementation(
+export async function getImplementation(
   provider: DiscoveryProvider,
   address: EthereumAddress,
 ) {
@@ -23,11 +23,14 @@ const ADMIN_SLOT = Bytes.fromHex(
   '0xb53127684a568b3173ae13b9f8a6016e243e63b6e8ee1178d6a717850b5d6103',
 )
 
-async function getAdmin(provider: DiscoveryProvider, address: EthereumAddress) {
+export async function getAdmin(
+  provider: DiscoveryProvider,
+  address: EthereumAddress,
+) {
   return bytes32ToAddress(await provider.getStorage(address, ADMIN_SLOT))
 }
 
-async function detect(
+export async function detectEip1967Proxy(
   provider: DiscoveryProvider,
   address: EthereumAddress,
 ): Promise<ProxyDetection | undefined> {
@@ -45,10 +48,4 @@ async function detect(
       admin,
     },
   }
-}
-
-export const Eip1967Proxy = {
-  getImplementation,
-  getAdmin,
-  detect,
 }
